CLN: Cleanup dataframe file structures
