+++ dask/dataframe/tests/test_dataframe.py
@@ -1,4 +1,3 @@
-from itertools import product
 from datetime import datetime
 from operator import getitem
 from distutils.version import LooseVersion
@@ -137,943 +136,6 @@ def test_set_index_raises_error_on_bad_input():
 
     assert raises(NotImplementedError, lambda: ddf.set_index(['a', 'b']))
 
-
-def test_split_apply_combine_on_series():
-    dsk = {('x', 0): pd.DataFrame({'a': [1, 2, 6], 'b': [4, 2, 7]},
-                                  index=[0, 1, 3]),
-           ('x', 1): pd.DataFrame({'a': [4, 4, 6], 'b': [3, 3, 1]},
-                                  index=[5, 6, 8]),
-           ('x', 2): pd.DataFrame({'a': [4, 3, 7], 'b': [1, 1, 3]},
-                                  index=[9, 9, 9])}
-    d = dd.DataFrame(dsk, 'x', ['a', 'b'], [0, 4, 9, 9])
-    full = d.compute()
-
-    for ddkey, pdkey in [('b', 'b'), (d.b, full.b),
-                         (d.b + 1, full.b + 1)]:
-        assert eq(d.groupby(ddkey).a.min(), full.groupby(pdkey).a.min())
-        assert eq(d.groupby(ddkey).a.max(), full.groupby(pdkey).a.max())
-        assert eq(d.groupby(ddkey).a.count(), full.groupby(pdkey).a.count())
-        assert eq(d.groupby(ddkey).a.mean(), full.groupby(pdkey).a.mean())
-        assert eq(d.groupby(ddkey).a.nunique(), full.groupby(pdkey).a.nunique())
-
-        assert eq(d.groupby(ddkey).sum(), full.groupby(pdkey).sum())
-        assert eq(d.groupby(ddkey).min(), full.groupby(pdkey).min())
-        assert eq(d.groupby(ddkey).max(), full.groupby(pdkey).max())
-        assert eq(d.groupby(ddkey).count(), full.groupby(pdkey).count())
-        assert eq(d.groupby(ddkey).mean(), full.groupby(pdkey).mean())
-
-    for ddkey, pdkey in [(d.b, full.b), (d.b + 1, full.b + 1)]:
-        assert eq(d.a.groupby(ddkey).sum(), full.a.groupby(pdkey).sum(), check_names=False)
-        assert eq(d.a.groupby(ddkey).max(), full.a.groupby(pdkey).max(), check_names=False)
-        assert eq(d.a.groupby(ddkey).count(), full.a.groupby(pdkey).count(), check_names=False)
-        assert eq(d.a.groupby(ddkey).mean(), full.a.groupby(pdkey).mean(), check_names=False)
-        assert eq(d.a.groupby(ddkey).nunique(), full.a.groupby(pdkey).nunique(), check_names=False)
-
-    for i in range(8):
-        assert eq(d.groupby(d.b > i).a.sum(), full.groupby(full.b > i).a.sum())
-        assert eq(d.groupby(d.b > i).a.min(), full.groupby(full.b > i).a.min())
-        assert eq(d.groupby(d.b > i).a.max(), full.groupby(full.b > i).a.max())
-        assert eq(d.groupby(d.b > i).a.count(), full.groupby(full.b > i).a.count())
-        assert eq(d.groupby(d.b > i).a.mean(), full.groupby(full.b > i).a.mean())
-        assert eq(d.groupby(d.b > i).a.nunique(), full.groupby(full.b > i).a.nunique())
-
-        assert eq(d.groupby(d.a > i).b.sum(), full.groupby(full.a > i).b.sum())
-        assert eq(d.groupby(d.a > i).b.min(), full.groupby(full.a > i).b.min())
-        assert eq(d.groupby(d.a > i).b.max(), full.groupby(full.a > i).b.max())
-        assert eq(d.groupby(d.a > i).b.count(), full.groupby(full.a > i).b.count())
-        assert eq(d.groupby(d.a > i).b.mean(), full.groupby(full.a > i).b.mean())
-        assert eq(d.groupby(d.a > i).b.nunique(), full.groupby(full.a > i).b.nunique())
-
-        assert eq(d.groupby(d.b > i).sum(), full.groupby(full.b > i).sum())
-        assert eq(d.groupby(d.b > i).min(), full.groupby(full.b > i).min())
-        assert eq(d.groupby(d.b > i).max(), full.groupby(full.b > i).max())
-        assert eq(d.groupby(d.b > i).count(), full.groupby(full.b > i).count())
-        assert eq(d.groupby(d.b > i).mean(), full.groupby(full.b > i).mean())
-
-        assert eq(d.groupby(d.a > i).sum(), full.groupby(full.a > i).sum())
-        assert eq(d.groupby(d.a > i).min(), full.groupby(full.a > i).min())
-        assert eq(d.groupby(d.a > i).max(), full.groupby(full.a > i).max())
-        assert eq(d.groupby(d.a > i).count(), full.groupby(full.a > i).count())
-        assert eq(d.groupby(d.a > i).mean(), full.groupby(full.a > i).mean())
-
-    for ddkey, pdkey in [('a', 'a'), (d.a, full.a),
-                         (d.a + 1, full.a + 1), (d.a > 3, full.a > 3)]:
-        assert eq(d.groupby(ddkey).b.sum(), full.groupby(pdkey).b.sum())
-        assert eq(d.groupby(ddkey).b.min(), full.groupby(pdkey).b.min())
-        assert eq(d.groupby(ddkey).b.max(), full.groupby(pdkey).b.max())
-        assert eq(d.groupby(ddkey).b.count(), full.groupby(pdkey).b.count())
-        assert eq(d.groupby(ddkey).b.mean(), full.groupby(pdkey).b.mean())
-        assert eq(d.groupby(ddkey).b.nunique(), full.groupby(pdkey).b.nunique())
-
-        assert eq(d.groupby(ddkey).sum(), full.groupby(pdkey).sum())
-        assert eq(d.groupby(ddkey).min(), full.groupby(pdkey).min())
-        assert eq(d.groupby(ddkey).max(), full.groupby(pdkey).max())
-        assert eq(d.groupby(ddkey).count(), full.groupby(pdkey).count())
-        assert eq(d.groupby(ddkey).mean(), full.groupby(pdkey).mean().astype(float))
-
-    assert sorted(d.groupby('b').a.sum().dask) == \
-           sorted(d.groupby('b').a.sum().dask)
-    assert sorted(d.groupby(d.a > 3).b.mean().dask) == \
-           sorted(d.groupby(d.a > 3).b.mean().dask)
-
-    # test raises with incorrect key
-    assert raises(KeyError, lambda: d.groupby('x'))
-    assert raises(KeyError, lambda: d.groupby(['a', 'x']))
-    assert raises(KeyError, lambda: d.groupby('a')['x'])
-    assert raises(KeyError, lambda: d.groupby('a')['b', 'x'])
-    assert raises(KeyError, lambda: d.groupby('a')[['b', 'x']])
-
-    # test graph node labels
-    assert_dask_graph(d.groupby('b').a.sum(), 'series-groupby-sum')
-    assert_dask_graph(d.groupby('b').a.min(), 'series-groupby-min')
-    assert_dask_graph(d.groupby('b').a.max(), 'series-groupby-max')
-    assert_dask_graph(d.groupby('b').a.count(), 'series-groupby-count')
-    # mean consists from sum and count operations
-    assert_dask_graph(d.groupby('b').a.mean(), 'series-groupby-sum')
-    assert_dask_graph(d.groupby('b').a.mean(), 'series-groupby-count')
-    assert_dask_graph(d.groupby('b').a.nunique(), 'series-groupby-nunique')
-
-    assert_dask_graph(d.groupby('b').sum(), 'dataframe-groupby-sum')
-    assert_dask_graph(d.groupby('b').min(), 'dataframe-groupby-min')
-    assert_dask_graph(d.groupby('b').max(), 'dataframe-groupby-max')
-    assert_dask_graph(d.groupby('b').count(), 'dataframe-groupby-count')
-    # mean consists from sum and count operations
-    assert_dask_graph(d.groupby('b').mean(), 'dataframe-groupby-sum')
-    assert_dask_graph(d.groupby('b').mean(), 'dataframe-groupby-count')
-
-
-def test_groupby_multilevel_getitem():
-    df = pd.DataFrame({'a': [1, 2, 3, 1, 2, 3],
-                       'b': [1, 2, 1, 4, 2, 1],
-                       'c': [1, 3, 2, 1, 1, 2],
-                       'd': [1, 2, 1, 1, 2, 2]})
-    ddf = dd.from_pandas(df, 2)
-
-    cases = [(ddf.groupby('a')['b'], df.groupby('a')['b']),
-             (ddf.groupby(['a', 'b']), df.groupby(['a', 'b'])),
-             (ddf.groupby(['a', 'b'])['c'], df.groupby(['a', 'b'])['c']),
-             (ddf.groupby('a')[['b', 'c']], df.groupby('a')[['b', 'c']]),
-             (ddf.groupby('a')[['b']], df.groupby('a')[['b']]),
-             (ddf.groupby(['a', 'b', 'c']), df.groupby(['a', 'b', 'c']))]
-
-    for d, p in cases:
-        assert isinstance(d, dd.core._GroupBy)
-        assert isinstance(p, pd.core.groupby.GroupBy)
-        assert eq(d.sum(), p.sum())
-        assert eq(d.min(), p.min())
-        assert eq(d.max(), p.max())
-        assert eq(d.count(), p.count())
-        assert eq(d.mean(), p.mean().astype(float))
-
-def test_groupby_get_group():
-    dsk = {('x', 0): pd.DataFrame({'a': [1, 2, 6], 'b': [4, 2, 7]},
-                                  index=[0, 1, 3]),
-           ('x', 1): pd.DataFrame({'a': [4, 2, 6], 'b': [3, 3, 1]},
-                                  index=[5, 6, 8]),
-           ('x', 2): pd.DataFrame({'a': [4, 3, 7], 'b': [1, 1, 3]},
-                                  index=[9, 9, 9])}
-    d = dd.DataFrame(dsk, 'x', ['a', 'b'], [0, 4, 9, 9])
-    full = d.compute()
-
-    for ddkey, pdkey in [('b', 'b'), (d.b, full.b),
-                         (d.b + 1, full.b + 1)]:
-        ddgrouped = d.groupby(ddkey)
-        pdgrouped = full.groupby(pdkey)
-        # DataFrame
-        assert eq(ddgrouped.get_group(2), pdgrouped.get_group(2))
-        assert eq(ddgrouped.get_group(3), pdgrouped.get_group(3))
-        # Series
-        assert eq(ddgrouped.a.get_group(3), pdgrouped.a.get_group(3))
-        assert eq(ddgrouped.a.get_group(2), pdgrouped.a.get_group(2))
-
-def test_arithmetics():
-    pdf2 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8],
-                         'b': [5, 6, 7, 8, 1, 2, 3, 4]})
-    pdf3 = pd.DataFrame({'a': [5, 6, 7, 8, 4, 3, 2, 1],
-                         'b': [2, 4, 5, 3, 4, 2, 1, 0]})
-    ddf2 = dd.from_pandas(pdf2, 3)
-    ddf3 = dd.from_pandas(pdf3, 2)
-
-    dsk4 = {('y', 0): pd.DataFrame({'a': [3, 2, 1], 'b': [7, 8, 9]},
-                                   index=[0, 1, 3]),
-            ('y', 1): pd.DataFrame({'a': [5, 2, 8], 'b': [4, 2, 3]},
-                                   index=[5, 6, 8]),
-            ('y', 2): pd.DataFrame({'a': [1, 4, 10], 'b': [1, 0, 5]},
-                                   index=[9, 9, 9])}
-    ddf4 = dd.DataFrame(dsk4, 'y', ['a', 'b'], [0, 4, 9, 9])
-    pdf4 =ddf4.compute()
-
-    # Arithmetics
-    cases = [(d, d, full, full),
-             (d, d.repartition([0, 1, 3, 6, 9]), full, full),
-             (ddf2, ddf3, pdf2, pdf3),
-             (ddf2.repartition([0, 3, 6, 7]), ddf3.repartition([0, 7]),
-              pdf2, pdf3),
-             (ddf2.repartition([0, 7]), ddf3.repartition([0, 2, 4, 5, 7]),
-              pdf2, pdf3),
-             (d, ddf4, full, pdf4),
-             (d, ddf4.repartition([0, 9]), full, pdf4),
-             (d.repartition([0, 3, 9]), ddf4.repartition([0, 5, 9]),
-              full, pdf4),
-             # dask + pandas
-             (d, pdf4, full, pdf4), (ddf2, pdf3, pdf2, pdf3)]
-
-    for (l, r, el, er) in cases:
-        check_series_arithmetics(l.a, r.b, el.a, er.b)
-        check_frame_arithmetics(l, r, el, er)
-
-    # different index, pandas raises ValueError in comparison ops
-
-    pdf5 = pd.DataFrame({'a': [3, 2, 1, 5, 2, 8, 1, 4, 10],
-                         'b': [7, 8, 9, 4, 2, 3, 1, 0, 5]},
-                        index=[0, 1, 3, 5, 6, 8, 9, 9, 9])
-    ddf5 = dd.from_pandas(pdf5, 2)
-
-    pdf6 = pd.DataFrame({'a': [3, 2, 1, 5, 2 ,8, 1, 4, 10],
-                         'b': [7, 8, 9, 5, 7, 8, 4, 2, 5]},
-                        index=[0, 1, 2, 3, 4, 5, 6, 7, 9])
-    ddf6 = dd.from_pandas(pdf6, 4)
-
-    pdf7 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8],
-                         'b': [5, 6, 7, 8, 1, 2, 3, 4]},
-                        index=list('aaabcdeh'))
-    pdf8 = pd.DataFrame({'a': [5, 6, 7, 8, 4, 3, 2, 1],
-                         'b': [2, 4, 5, 3, 4, 2, 1, 0]},
-                        index=list('abcdefgh'))
-    ddf7 = dd.from_pandas(pdf7, 3)
-    ddf8 = dd.from_pandas(pdf8, 4)
-
-    pdf9 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8],
-                         'b': [5, 6, 7, 8, 1, 2, 3, 4],
-                         'c': [5, 6, 7, 8, 1, 2, 3, 4]},
-                        index=list('aaabcdeh'))
-    pdf10 = pd.DataFrame({'b': [5, 6, 7, 8, 4, 3, 2, 1],
-                          'c': [2, 4, 5, 3, 4, 2, 1, 0],
-                          'd': [2, 4, 5, 3, 4, 2, 1, 0]},
-                        index=list('abcdefgh'))
-    ddf9 = dd.from_pandas(pdf9, 3)
-    ddf10 = dd.from_pandas(pdf10, 4)
-
-    # Arithmetics with different index
-    cases = [(ddf5, ddf6, pdf5, pdf6),
-             (ddf5.repartition([0, 9]), ddf6, pdf5, pdf6),
-             (ddf5.repartition([0, 5, 9]), ddf6.repartition([0, 7, 9]),
-              pdf5, pdf6),
-             (ddf7, ddf8, pdf7, pdf8),
-             (ddf7.repartition(['a', 'c', 'h']), ddf8.repartition(['a', 'h']),
-              pdf7, pdf8),
-             (ddf7.repartition(['a', 'b', 'e', 'h']),
-              ddf8.repartition(['a', 'e', 'h']), pdf7, pdf8),
-             (ddf9, ddf10, pdf9, pdf10),
-             (ddf9.repartition(['a', 'c', 'h']), ddf10.repartition(['a', 'h']),
-              pdf9, pdf10),
-             # dask + pandas
-             (ddf5, pdf6, pdf5, pdf6), (ddf7, pdf8, pdf7, pdf8),
-             (ddf9, pdf10, pdf9, pdf10)]
-
-    for (l, r, el, er) in cases:
-        check_series_arithmetics(l.a, r.b, el.a, er.b,
-                                 allow_comparison_ops=False)
-        check_frame_arithmetics(l, r, el, er,
-                                allow_comparison_ops=False)
-
-def test_arithmetics_different_index():
-
-    # index are different, but overwraps
-    pdf1 = pd.DataFrame({'a': [1, 2, 3, 4, 5], 'b': [3, 5, 2, 5, 7]},
-                        index=[1, 2, 3, 4, 5])
-    ddf1 = dd.from_pandas(pdf1, 2)
-    pdf2 = pd.DataFrame({'a': [3, 2, 6, 7, 8], 'b': [9, 4, 2, 6, 2]},
-                        index=[3, 4, 5, 6, 7])
-    ddf2 = dd.from_pandas(pdf2, 2)
-
-    # index are not overwrapped
-    pdf3 = pd.DataFrame({'a': [1, 2, 3, 4, 5], 'b': [3, 5, 2, 5, 7]},
-                        index=[1, 2, 3, 4, 5])
-    ddf3 = dd.from_pandas(pdf3, 2)
-    pdf4 = pd.DataFrame({'a': [3, 2, 6, 7, 8], 'b': [9, 4, 2, 6, 2]},
-                        index=[10, 11, 12, 13, 14])
-    ddf4 = dd.from_pandas(pdf4, 2)
-
-    # index is included in another
-    pdf5 = pd.DataFrame({'a': [1, 2, 3, 4, 5], 'b': [3, 5, 2, 5, 7]},
-                        index=[1, 3, 5, 7, 9])
-    ddf5 = dd.from_pandas(pdf5, 2)
-    pdf6 = pd.DataFrame({'a': [3, 2, 6, 7, 8], 'b': [9, 4, 2, 6, 2]},
-                        index=[2, 3, 4, 5, 6])
-    ddf6 = dd.from_pandas(pdf6, 2)
-
-    cases = [(ddf1, ddf2, pdf1, pdf2),
-             (ddf2, ddf1, pdf2, pdf1),
-             (ddf1.repartition([1, 3, 5]), ddf2.repartition([3, 4, 7]),
-              pdf1, pdf2),
-             (ddf2.repartition([3, 4, 5, 7]), ddf1.repartition([1, 2, 4, 5]),
-              pdf2, pdf1),
-             (ddf3, ddf4, pdf3, pdf4),
-             (ddf4, ddf3, pdf4, pdf3),
-             (ddf3.repartition([1, 2, 3, 4, 5]),
-              ddf4.repartition([10, 11, 12, 13, 14]), pdf3, pdf4),
-             (ddf4.repartition([10, 14]), ddf3.repartition([1, 3, 4, 5]),
-              pdf4, pdf3),
-             (ddf5, ddf6, pdf5, pdf6),
-             (ddf6, ddf5, pdf6, pdf5),
-             (ddf5.repartition([1, 7, 8, 9]), ddf6.repartition([2, 3, 4, 6]),
-              pdf5, pdf6),
-             (ddf6.repartition([2, 6]), ddf5.repartition([1, 3, 7, 9]),
-              pdf6, pdf5),
-             # dask + pandas
-             (ddf1, pdf2, pdf1, pdf2), (ddf2, pdf1, pdf2, pdf1),
-             (ddf3, pdf4, pdf3, pdf4), (ddf4, pdf3, pdf4, pdf3),
-             (ddf5, pdf6, pdf5, pdf6), (ddf6, pdf5, pdf6, pdf5)]
-
-    for (l, r, el, er) in cases:
-        check_series_arithmetics(l.a, r.b, el.a, er.b,
-                                 allow_comparison_ops=False)
-        check_frame_arithmetics(l, r, el, er,
-                                allow_comparison_ops=False)
-
-    pdf7 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8],
-                          'b': [5, 6, 7, 8, 1, 2, 3, 4]},
-                         index=[0, 2, 4, 8, 9, 10, 11, 13])
-    pdf8 = pd.DataFrame({'a': [5, 6, 7, 8, 4, 3, 2, 1],
-                          'b': [2, 4, 5, 3, 4, 2, 1, 0]},
-                         index=[1, 3, 4, 8, 9, 11, 12, 13])
-    ddf7 = dd.from_pandas(pdf7, 3)
-    ddf8 = dd.from_pandas(pdf8, 2)
-
-    pdf9 = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7, 8],
-                         'b': [5, 6, 7, 8, 1, 2, 3, 4]},
-                        index=[0, 2, 4, 8, 9, 10, 11, 13])
-    pdf10 = pd.DataFrame({'a': [5, 6, 7, 8, 4, 3, 2, 1],
-                          'b': [2, 4, 5, 3, 4, 2, 1, 0]},
-                         index=[0, 3, 4, 8, 9, 11, 12, 13])
-    ddf9 = dd.from_pandas(pdf9, 3)
-    ddf10 = dd.from_pandas(pdf10, 2)
-
-    cases = [(ddf7, ddf8, pdf7, pdf8),
-             (ddf8, ddf7, pdf8, pdf7),
-             (ddf7.repartition([0, 13]),
-              ddf8.repartition([0, 4, 11, 14], force=True),
-              pdf7, pdf8),
-             (ddf8.repartition([-5, 10, 15], force=True),
-              ddf7.repartition([-1, 4, 11, 14], force=True), pdf8, pdf7),
-             (ddf7.repartition([0, 8, 12, 13]),
-              ddf8.repartition([0, 2, 8, 12, 13], force=True), pdf7, pdf8),
-             (ddf8.repartition([-5, 0, 10, 20], force=True),
-              ddf7.repartition([-1, 4, 11, 13], force=True), pdf8, pdf7),
-             (ddf9, ddf10, pdf9, pdf10),
-             (ddf10, ddf9, pdf10, pdf9),
-             # dask + pandas
-             (ddf7, pdf8, pdf7, pdf8), (ddf8, pdf7, pdf8, pdf7),
-             (ddf9, pdf10, pdf9, pdf10), (ddf10, pdf9, pdf10, pdf9)]
-
-    for (l, r, el, er) in cases:
-        check_series_arithmetics(l.a, r.b, el.a, er.b,
-                                 allow_comparison_ops=False)
-        check_frame_arithmetics(l, r, el, er,
-                                allow_comparison_ops=False)
-
-
-def check_series_arithmetics(l, r, el, er, allow_comparison_ops=True):
-    assert isinstance(l, dd.Series)
-    assert isinstance(r, (dd.Series, pd.Series))
-    assert isinstance(el, pd.Series)
-    assert isinstance(er, pd.Series)
-
-    # l, r may be repartitioned, test whether repartition keeps original data
-    assert eq(l, el)
-    assert eq(r, er)
-
-    assert eq(l + r, el + er)
-    assert eq(l * r, el * er)
-    assert eq(l - r, el - er)
-    assert eq(l / r, el / er)
-    assert eq(l // r, el // er)
-    assert eq(l ** r, el ** er)
-    assert eq(l % r, el % er)
-
-    if allow_comparison_ops:
-        # comparison is allowed if data have same index
-        assert eq(l & r, el & er)
-        assert eq(l | r, el | er)
-        assert eq(l ^ r, el ^ er)
-        assert eq(l > r, el > er)
-        assert eq(l < r, el < er)
-        assert eq(l >= r, el >= er)
-        assert eq(l <= r, el <= er)
-        assert eq(l == r, el == er)
-        assert eq(l != r, el != er)
-
-    assert eq(l + 2, el + 2)
-    assert eq(l * 2, el * 2)
-    assert eq(l - 2, el - 2)
-    assert eq(l / 2, el / 2)
-    assert eq(l & True, el & True)
-    assert eq(l | True, el | True)
-    assert eq(l ^ True, el ^ True)
-    assert eq(l // 2, el // 2)
-    assert eq(l ** 2, el ** 2)
-    assert eq(l % 2, el % 2)
-    assert eq(l > 2, el > 2)
-    assert eq(l < 2, el < 2)
-    assert eq(l >= 2, el >= 2)
-    assert eq(l <= 2, el <= 2)
-    assert eq(l == 2, el == 2)
-    assert eq(l != 2, el != 2)
-
-    assert eq(2 + r, 2 + er)
-    assert eq(2 * r, 2 * er)
-    assert eq(2 - r, 2 - er)
-    assert eq(2 / r, 2 / er)
-    assert eq(True & r, True & er)
-    assert eq(True | r, True | er)
-    assert eq(True ^ r, True ^ er)
-    assert eq(2 // r, 2 // er)
-    assert eq(2 ** r, 2 ** er)
-    assert eq(2 % r, 2 % er)
-    assert eq(2 > r, 2 > er)
-    assert eq(2 < r, 2 < er)
-    assert eq(2 >= r, 2 >= er)
-    assert eq(2 <= r, 2 <= er)
-    assert eq(2 == r, 2 == er)
-    assert eq(2 != r, 2 != er)
-
-    assert eq(-l, -el)
-    assert eq(abs(l), abs(el))
-
-    if allow_comparison_ops:
-        # comparison is allowed if data have same index
-        assert eq(~(l == r), ~(el == er))
-
-
-def check_frame_arithmetics(l, r, el, er, allow_comparison_ops=True):
-    assert isinstance(l, dd.DataFrame)
-    assert isinstance(r, (dd.DataFrame, pd.DataFrame))
-    assert isinstance(el, pd.DataFrame)
-    assert isinstance(er, pd.DataFrame)
-    # l, r may be repartitioned, test whether repartition keeps original data
-    assert eq(l, el)
-    assert eq(r, er)
-
-    assert eq(l + r, el + er)
-    assert eq(l * r, el * er)
-    assert eq(l - r, el - er)
-    assert eq(l / r, el / er)
-    assert eq(l // r, el // er)
-    assert eq(l ** r, el ** er)
-    assert eq(l % r, el % er)
-
-    if allow_comparison_ops:
-        # comparison is allowed if data have same index
-        assert eq(l & r, el & er)
-        assert eq(l | r, el | er)
-        assert eq(l ^ r, el ^ er)
-        assert eq(l > r, el > er)
-        assert eq(l < r, el < er)
-        assert eq(l >= r, el >= er)
-        assert eq(l <= r, el <= er)
-        assert eq(l == r, el == er)
-        assert eq(l != r, el != er)
-
-    assert eq(l + 2, el + 2)
-    assert eq(l * 2, el * 2)
-    assert eq(l - 2, el - 2)
-    assert eq(l / 2, el / 2)
-    assert eq(l & True, el & True)
-    assert eq(l | True, el | True)
-    assert eq(l ^ True, el ^ True)
-    assert eq(l // 2, el // 2)
-    assert eq(l ** 2, el ** 2)
-    assert eq(l % 2, el % 2)
-    assert eq(l > 2, el > 2)
-    assert eq(l < 2, el < 2)
-    assert eq(l >= 2, el >= 2)
-    assert eq(l <= 2, el <= 2)
-    assert eq(l == 2, el == 2)
-    assert eq(l != 2, el != 2)
-
-    assert eq(2 + l, 2 + el)
-    assert eq(2 * l, 2 * el)
-    assert eq(2 - l, 2 - el)
-    assert eq(2 / l, 2 / el)
-    assert eq(True & l, True & el)
-    assert eq(True | l, True | el)
-    assert eq(True ^ l, True ^ el)
-    assert eq(2 // l, 2 // el)
-    assert eq(2 ** l, 2 ** el)
-    assert eq(2 % l, 2 % el)
-    assert eq(2 > l, 2 > el)
-    assert eq(2 < l, 2 < el)
-    assert eq(2 >= l, 2 >= el)
-    assert eq(2 <= l, 2 <= el)
-    assert eq(2 == l, 2 == el)
-    assert eq(2 != l, 2 != el)
-
-    assert eq(-l, -el)
-    assert eq(abs(l), abs(el))
-
-    if allow_comparison_ops:
-        # comparison is allowed if data have same index
-        assert eq(~(l == r), ~(el == er))
-
-
-def test_scalar_arithmetics():
-    l = dd.core.Scalar({('l', 0): 10}, 'l')
-    r = dd.core.Scalar({('r', 0): 4}, 'r')
-    el = 10
-    er = 4
-
-    assert isinstance(l, dd.core.Scalar)
-    assert isinstance(r, dd.core.Scalar)
-
-    # l, r may be repartitioned, test whether repartition keeps original data
-    assert eq(l, el)
-    assert eq(r, er)
-
-    assert eq(l + r, el + er)
-    assert eq(l * r, el * er)
-    assert eq(l - r, el - er)
-    assert eq(l / r, el / er)
-    assert eq(l // r, el // er)
-    assert eq(l ** r, el ** er)
-    assert eq(l % r, el % er)
-
-    assert eq(l & r, el & er)
-    assert eq(l | r, el | er)
-    assert eq(l ^ r, el ^ er)
-    assert eq(l > r, el > er)
-    assert eq(l < r, el < er)
-    assert eq(l >= r, el >= er)
-    assert eq(l <= r, el <= er)
-    assert eq(l == r, el == er)
-    assert eq(l != r, el != er)
-
-    assert eq(l + 2, el + 2)
-    assert eq(l * 2, el * 2)
-    assert eq(l - 2, el - 2)
-    assert eq(l / 2, el / 2)
-    assert eq(l & True, el & True)
-    assert eq(l | True, el | True)
-    assert eq(l ^ True, el ^ True)
-    assert eq(l // 2, el // 2)
-    assert eq(l ** 2, el ** 2)
-    assert eq(l % 2, el % 2)
-    assert eq(l > 2, el > 2)
-    assert eq(l < 2, el < 2)
-    assert eq(l >= 2, el >= 2)
-    assert eq(l <= 2, el <= 2)
-    assert eq(l == 2, el == 2)
-    assert eq(l != 2, el != 2)
-
-    assert eq(2 + r, 2 + er)
-    assert eq(2 * r, 2 * er)
-    assert eq(2 - r, 2 - er)
-    assert eq(2 / r, 2 / er)
-    assert eq(True & r, True & er)
-    assert eq(True | r, True | er)
-    assert eq(True ^ r, True ^ er)
-    assert eq(2 // r, 2 // er)
-    assert eq(2 ** r, 2 ** er)
-    assert eq(2 % r, 2 % er)
-    assert eq(2 > r, 2 > er)
-    assert eq(2 < r, 2 < er)
-    assert eq(2 >= r, 2 >= er)
-    assert eq(2 <= r, 2 <= er)
-    assert eq(2 == r, 2 == er)
-    assert eq(2 != r, 2 != er)
-
-    assert eq(-l, -el)
-    assert eq(abs(l), abs(el))
-
-    assert eq(~(l == r), ~(el == er))
-
-
-def test_scalar_arithmetics_with_dask_instances():
-    s = dd.core.Scalar({('s', 0): 10}, 's')
-    e = 10
-
-    pds = pd.Series([1, 2, 3, 4, 5, 6, 7])
-    dds = dd.from_pandas(pds, 2)
-
-    pdf = pd.DataFrame({'a': [1, 2, 3, 4, 5, 6, 7],
-                        'b': [7, 6, 5, 4, 3, 2, 1]})
-    ddf = dd.from_pandas(pdf, 2)
-
-    # pandas Series
-    result = pds + s   # this result pd.Series (automatically computed)
-    assert isinstance(result, pd.Series)
-    assert eq(result, pds + e)
-
-    result = s + pds   # this result dd.Series
-    assert isinstance(result, dd.Series)
-    assert eq(result, pds + e)
-
-    # dask Series
-    result = dds + s   # this result dd.Series
-    assert isinstance(result, dd.Series)
-    assert eq(result, pds + e)
-
-    result = s + dds   # this result dd.Series
-    assert isinstance(result, dd.Series)
-    assert eq(result, pds + e)
-
-
-    # pandas DataFrame
-    result = pdf + s   # this result pd.DataFrame (automatically computed)
-    assert isinstance(result, pd.DataFrame)
-    assert eq(result, pdf + e)
-
-    result = s + pdf   # this result dd.DataFrame
-    assert isinstance(result, dd.DataFrame)
-    assert eq(result, pdf + e)
-
-    # dask DataFrame
-    result = ddf + s   # this result dd.DataFrame
-    assert isinstance(result, dd.DataFrame)
-    assert eq(result, pdf + e)
-
-    result = s + ddf   # this result dd.DataFrame
-    assert isinstance(result, dd.DataFrame)
-    assert eq(result, pdf + e)
-
-
-def test_frame_series_arithmetic_methods():
-
-    pdf1 = pd.DataFrame({'A': np.arange(10),
-                         'B': [np.nan, 1, 2, 3, 4] * 2,
-                         'C': [np.nan] * 10,
-                         'D': np.arange(10)},
-                        index=list('abcdefghij'), columns=list('ABCD'))
-    pdf2 = pd.DataFrame(np.random.randn(10, 4),
-                        index=list('abcdefghjk'), columns=list('ABCX'))
-    ps1 = pdf1.A
-    ps2 = pdf2.A
-
-    ddf1 = dd.from_pandas(pdf1, 2)
-    ddf2 = dd.from_pandas(pdf2, 2)
-    ds1 = ddf1.A
-    ds2 = ddf2.A
-
-    s = dd.core.Scalar({('s', 0): 4}, 's')
-
-    for l, r, el, er in [(ddf1, ddf2, pdf1, pdf2), (ds1, ds2, ps1, ps2),
-                         (ddf1.repartition(['a', 'f', 'j']), ddf2, pdf1, pdf2),
-                         (ds1.repartition(['a', 'b', 'f', 'j']), ds2, ps1, ps2),
-                         (ddf1, ddf2.repartition(['a', 'k']), pdf1, pdf2),
-                         (ds1, ds2.repartition(['a', 'b', 'd', 'h', 'k']), ps1, ps2),
-                         (ddf1, 3, pdf1, 3), (ds1, 3, ps1, 3),
-                         (ddf1, s, pdf1, 4), (ds1, s, ps1, 4)]:
-        # l, r may be repartitioned, test whether repartition keeps original data
-        assert eq(l, el)
-        assert eq(r, er)
-
-        assert eq(l.add(r, fill_value=0), el.add(er, fill_value=0))
-        assert eq(l.sub(r, fill_value=0), el.sub(er, fill_value=0))
-        assert eq(l.mul(r, fill_value=0), el.mul(er, fill_value=0))
-        assert eq(l.div(r, fill_value=0), el.div(er, fill_value=0))
-        assert eq(l.truediv(r, fill_value=0), el.truediv(er, fill_value=0))
-        assert eq(l.floordiv(r, fill_value=1), el.floordiv(er, fill_value=1))
-        assert eq(l.mod(r, fill_value=0), el.mod(er, fill_value=0))
-        assert eq(l.pow(r, fill_value=0), el.pow(er, fill_value=0))
-
-        assert eq(l.radd(r, fill_value=0), el.radd(er, fill_value=0))
-        assert eq(l.rsub(r, fill_value=0), el.rsub(er, fill_value=0))
-        assert eq(l.rmul(r, fill_value=0), el.rmul(er, fill_value=0))
-        assert eq(l.rdiv(r, fill_value=0), el.rdiv(er, fill_value=0))
-        assert eq(l.rtruediv(r, fill_value=0), el.rtruediv(er, fill_value=0))
-        assert eq(l.rfloordiv(r, fill_value=1), el.rfloordiv(er, fill_value=1))
-        assert eq(l.rmod(r, fill_value=0), el.rmod(er, fill_value=0))
-        assert eq(l.rpow(r, fill_value=0), el.rpow(er, fill_value=0))
-
-    for l, r, el, er in [(ddf1, ds2, pdf1, ps2), (ddf1, ddf2.X, pdf1, pdf2.X)]:
-        assert eq(l, el)
-        assert eq(r, er)
-
-        # must specify axis=0 to add Series to each column
-        # axis=1 is not supported (add to each row)
-        assert eq(l.add(r, axis=0), el.add(er, axis=0))
-        assert eq(l.sub(r, axis=0), el.sub(er, axis=0))
-        assert eq(l.mul(r, axis=0), el.mul(er, axis=0))
-        assert eq(l.div(r, axis=0), el.div(er, axis=0))
-        assert eq(l.truediv(r, axis=0), el.truediv(er, axis=0))
-        assert eq(l.floordiv(r, axis=0), el.floordiv(er, axis=0))
-        assert eq(l.mod(r, axis=0), el.mod(er, axis=0))
-        assert eq(l.pow(r, axis=0), el.pow(er, axis=0))
-
-        assert eq(l.radd(r, axis=0), el.radd(er, axis=0))
-        assert eq(l.rsub(r, axis=0), el.rsub(er, axis=0))
-        assert eq(l.rmul(r, axis=0), el.rmul(er, axis=0))
-        assert eq(l.rdiv(r, axis=0), el.rdiv(er, axis=0))
-        assert eq(l.rtruediv(r, axis=0), el.rtruediv(er, axis=0))
-        assert eq(l.rfloordiv(r, axis=0), el.rfloordiv(er, axis=0))
-        assert eq(l.rmod(r, axis=0), el.rmod(er, axis=0))
-        assert eq(l.rpow(r, axis=0), el.rpow(er, axis=0))
-
-        assert raises(ValueError, lambda: l.add(r, axis=1))
-
-    for l, r, el, er in [(ddf1, pdf2, pdf1, pdf2), (ddf1, ps2, pdf1, ps2)]:
-        assert eq(l, el)
-        assert eq(r, er)
-
-        for axis in [0, 1, 'index', 'columns']:
-            assert eq(l.add(r, axis=axis), el.add(er, axis=axis))
-            assert eq(l.sub(r, axis=axis), el.sub(er, axis=axis))
-            assert eq(l.mul(r, axis=axis), el.mul(er, axis=axis))
-            assert eq(l.div(r, axis=axis), el.div(er, axis=axis))
-            assert eq(l.truediv(r, axis=axis), el.truediv(er, axis=axis))
-            assert eq(l.floordiv(r, axis=axis), el.floordiv(er, axis=axis))
-            assert eq(l.mod(r, axis=axis), el.mod(er, axis=axis))
-            assert eq(l.pow(r, axis=axis), el.pow(er, axis=axis))
-
-            assert eq(l.radd(r, axis=axis), el.radd(er, axis=axis))
-            assert eq(l.rsub(r, axis=axis), el.rsub(er, axis=axis))
-            assert eq(l.rmul(r, axis=axis), el.rmul(er, axis=axis))
-            assert eq(l.rdiv(r, axis=axis), el.rdiv(er, axis=axis))
-            assert eq(l.rtruediv(r, axis=axis), el.rtruediv(er, axis=axis))
-            assert eq(l.rfloordiv(r, axis=axis), el.rfloordiv(er, axis=axis))
-            assert eq(l.rmod(r, axis=axis), el.rmod(er, axis=axis))
-            assert eq(l.rpow(r, axis=axis), el.rpow(er, axis=axis))
-
-
-def test_reductions():
-    nans1 = pd.Series([1] + [np.nan] * 4 + [2] + [np.nan] * 3)
-    nands1 = dd.from_pandas(nans1, 2)
-    nans2 = pd.Series([1] + [np.nan] * 8)
-    nands2 = dd.from_pandas(nans2, 2)
-    nans3 = pd.Series([np.nan] * 9)
-    nands3 = dd.from_pandas(nans3, 2)
-
-    bools = pd.Series([True, False, True, False, True], dtype=bool)
-    boolds = dd.from_pandas(bools, 2)
-
-    for dds, pds in [(d.b, full.b), (d.a, full.a),
-                     (d['a'], full['a']), (d['b'], full['b']),
-                     (nands1, nans1), (nands2, nans2), (nands3, nans3),
-                     (boolds, bools)]:
-        assert isinstance(dds, dd.Series)
-        assert isinstance(pds, pd.Series)
-        assert eq(dds.sum(), pds.sum())
-        assert eq(dds.min(), pds.min())
-        assert eq(dds.max(), pds.max())
-        assert eq(dds.count(), pds.count())
-        assert eq(dds.std(), pds.std())
-        assert eq(dds.var(), pds.var())
-        assert eq(dds.std(ddof=0), pds.std(ddof=0))
-        assert eq(dds.var(ddof=0), pds.var(ddof=0))
-        assert eq(dds.mean(), pds.mean())
-        assert eq(dds.nunique(), pds.nunique())
-        assert eq(dds.nbytes, pds.nbytes)
-
-        assert eq(dds.sum(skipna=False), pds.sum(skipna=False))
-        assert eq(dds.min(skipna=False), pds.min(skipna=False))
-        assert eq(dds.max(skipna=False), pds.max(skipna=False))
-        assert eq(dds.std(skipna=False), pds.std(skipna=False))
-        assert eq(dds.var(skipna=False), pds.var(skipna=False))
-        assert eq(dds.std(skipna=False, ddof=0), pds.std(skipna=False, ddof=0))
-        assert eq(dds.var(skipna=False, ddof=0), pds.var(skipna=False, ddof=0))
-        assert eq(dds.mean(skipna=False), pds.mean(skipna=False))
-
-    assert_dask_graph(d.b.sum(), 'series-sum')
-    assert_dask_graph(d.b.min(), 'series-min')
-    assert_dask_graph(d.b.max(), 'series-max')
-    assert_dask_graph(d.b.count(), 'series-count')
-    assert_dask_graph(d.b.std(), 'series-std(ddof=1)')
-    assert_dask_graph(d.b.var(), 'series-var(ddof=1)')
-    assert_dask_graph(d.b.std(ddof=0), 'series-std(ddof=0)')
-    assert_dask_graph(d.b.var(ddof=0), 'series-var(ddof=0)')
-    assert_dask_graph(d.b.mean(), 'series-mean')
-    # nunique is performed using drop-duplicates
-    assert_dask_graph(d.b.nunique(), 'drop-duplicates')
-
-
-def test_reduction_series_invalid_axis():
-    for axis in [1, 'columns']:
-        for s in [d.a, full.a]: # both must behave the same
-            assert raises(ValueError, lambda: s.sum(axis=axis))
-            assert raises(ValueError, lambda: s.min(axis=axis))
-            assert raises(ValueError, lambda: s.max(axis=axis))
-            # only count doesn't have axis keyword
-            assert raises(TypeError, lambda: s.count(axis=axis))
-            assert raises(ValueError, lambda: s.std(axis=axis))
-            assert raises(ValueError, lambda: s.var(axis=axis))
-            assert raises(ValueError, lambda: s.mean(axis=axis))
-
-
-def test_reductions_non_numeric_dtypes():
-    # test non-numric blocks
-
-    def check_raises(d, p, func):
-        assert raises((TypeError, ValueError),
-                      lambda: getattr(d, func)().compute())
-        assert raises((TypeError, ValueError),
-                      lambda: getattr(p, func)())
-
-    pds = pd.Series(['a', 'b', 'c', 'd', 'e'])
-    dds = dd.from_pandas(pds, 2)
-    assert eq(dds.sum(), pds.sum())
-    assert eq(dds.min(), pds.min())
-    assert eq(dds.max(), pds.max())
-    assert eq(dds.count(), pds.count())
-    check_raises(dds, pds, 'std')
-    check_raises(dds, pds, 'var')
-    check_raises(dds, pds, 'mean')
-    assert eq(dds.nunique(), pds.nunique())
-
-    for pds in [pd.Series(pd.Categorical([1, 2, 3, 4, 5], ordered=True)),
-                pd.Series(pd.Categorical(list('abcde'), ordered=True)),
-                pd.Series(pd.date_range('2011-01-01', freq='D', periods=5))]:
-        dds = dd.from_pandas(pds, 2)
-
-        check_raises(dds, pds, 'sum')
-        assert eq(dds.min(), pds.min())
-        assert eq(dds.max(), pds.max())
-        assert eq(dds.count(), pds.count())
-        check_raises(dds, pds, 'std')
-        check_raises(dds, pds, 'var')
-        check_raises(dds, pds, 'mean')
-        assert eq(dds.nunique(), pds.nunique())
-
-
-    pds= pd.Series(pd.timedelta_range('1 days', freq='D', periods=5))
-    dds = dd.from_pandas(pds, 2)
-    assert eq(dds.sum(), pds.sum())
-    assert eq(dds.min(), pds.min())
-    assert eq(dds.max(), pds.max())
-    assert eq(dds.count(), pds.count())
-
-    # ToDo: pandas supports timedelta std, otherwise dask raises:
-    # incompatible type for a datetime/timedelta operation [__pow__]
-    # assert eq(dds.std(), pds.std())
-    # assert eq(dds.var(), pds.var())
-
-    # ToDo: pandas supports timedelta std, otherwise dask raises:
-    # TypeError: unsupported operand type(s) for *: 'float' and 'Timedelta'
-    # assert eq(dds.mean(), pds.mean())
-
-    assert eq(dds.nunique(), pds.nunique())
-
-
-def test_reductions_frame():
-    assert eq(d.sum(), full.sum())
-    assert eq(d.min(), full.min())
-    assert eq(d.max(), full.max())
-    assert eq(d.count(), full.count())
-    assert eq(d.std(), full.std())
-    assert eq(d.var(), full.var())
-    assert eq(d.std(ddof=0), full.std(ddof=0))
-    assert eq(d.var(ddof=0), full.var(ddof=0))
-    assert eq(d.mean(), full.mean())
-
-    for axis in [0, 1, 'index', 'columns']:
-        assert eq(d.sum(axis=axis), full.sum(axis=axis))
-        assert eq(d.min(axis=axis), full.min(axis=axis))
-        assert eq(d.max(axis=axis), full.max(axis=axis))
-        assert eq(d.count(axis=axis), full.count(axis=axis))
-        assert eq(d.std(axis=axis), full.std(axis=axis))
-        assert eq(d.var(axis=axis), full.var(axis=axis))
-        assert eq(d.std(axis=axis, ddof=0), full.std(axis=axis, ddof=0))
-        assert eq(d.var(axis=axis, ddof=0), full.var(axis=axis, ddof=0))
-        assert eq(d.mean(axis=axis), full.mean(axis=axis))
-
-    assert raises(ValueError, lambda: d.sum(axis='incorrect').compute())
-
-    # axis=0
-    assert_dask_graph(d.sum(), 'dataframe-sum')
-    assert_dask_graph(d.min(), 'dataframe-min')
-    assert_dask_graph(d.max(), 'dataframe-max')
-    assert_dask_graph(d.count(), 'dataframe-count')
-    # std, var, mean consists from sum and count operations
-    assert_dask_graph(d.std(), 'dataframe-sum')
-    assert_dask_graph(d.std(), 'dataframe-count')
-    assert_dask_graph(d.var(), 'dataframe-sum')
-    assert_dask_graph(d.var(), 'dataframe-count')
-    assert_dask_graph(d.mean(), 'dataframe-sum')
-    assert_dask_graph(d.mean(), 'dataframe-count')
-
-    # axis=1
-    assert_dask_graph(d.sum(axis=1), 'dataframe-sum(axis=1)')
-    assert_dask_graph(d.min(axis=1), 'dataframe-min(axis=1)')
-    assert_dask_graph(d.max(axis=1), 'dataframe-max(axis=1)')
-    assert_dask_graph(d.count(axis=1), 'dataframe-count(axis=1)')
-    assert_dask_graph(d.std(axis=1), 'dataframe-std(axis=1, ddof=1)')
-    assert_dask_graph(d.var(axis=1), 'dataframe-var(axis=1, ddof=1)')
-    assert_dask_graph(d.mean(axis=1), 'dataframe-mean(axis=1)')
-
-def test_reductions_frame_dtypes():
-    df = pd.DataFrame({'int': [1, 2, 3, 4, 5, 6, 7, 8],
-                       'float': [1., 2., 3., 4., np.nan, 6., 7., 8.],
-                       'dt': [pd.NaT] + [datetime(2011, i, 1) for i in range(1, 8)],
-                       'str': list('abcdefgh')})
-    ddf = dd.from_pandas(df, 3)
-    assert eq(df.sum(), ddf.sum())
-    assert eq(df.min(), ddf.min())
-    assert eq(df.max(), ddf.max())
-    assert eq(df.count(), ddf.count())
-    assert eq(df.std(), ddf.std())
-    assert eq(df.var(), ddf.var())
-    assert eq(df.std(ddof=0), ddf.std(ddof=0))
-    assert eq(df.var(ddof=0), ddf.var(ddof=0))
-    assert eq(df.mean(), ddf.mean())
-
-    assert eq(df._get_numeric_data(), ddf._get_numeric_data())
-
-    numerics = ddf[['int', 'float']]
-    assert numerics._get_numeric_data().dask == numerics.dask
-
-def test_reductions_frame_nan():
-    df = pd.DataFrame({'a': [1, 2, np.nan, 4, 5, 6, 7, 8],
-                       'b': [1, 2, np.nan, np.nan, np.nan, 5, np.nan, np.nan],
-                       'c': [np.nan] * 8})
-    ddf = dd.from_pandas(df, 3)
-    assert eq(df.sum(), ddf.sum())
-    assert eq(df.min(), ddf.min())
-    assert eq(df.max(), ddf.max())
-    assert eq(df.count(), ddf.count())
-    assert eq(df.std(), ddf.std())
-    assert eq(df.var(), ddf.var())
-    assert eq(df.std(ddof=0), ddf.std(ddof=0))
-    assert eq(df.var(ddof=0), ddf.var(ddof=0))
-    assert eq(df.mean(), ddf.mean())
-
-    assert eq(df.sum(skipna=False), ddf.sum(skipna=False))
-    assert eq(df.min(skipna=False), ddf.min(skipna=False))
-    assert eq(df.max(skipna=False), ddf.max(skipna=False))
-    assert eq(df.std(skipna=False), ddf.std(skipna=False))
-    assert eq(df.var(skipna=False), ddf.var(skipna=False))
-    assert eq(df.std(skipna=False, ddof=0), ddf.std(skipna=False, ddof=0))
-    assert eq(df.var(skipna=False, ddof=0), ddf.var(skipna=False, ddof=0))
-    assert eq(df.mean(skipna=False), ddf.mean(skipna=False))
-
-    assert eq(df.sum(axis=1, skipna=False), ddf.sum(axis=1, skipna=False))
-    assert eq(df.min(axis=1, skipna=False), ddf.min(axis=1, skipna=False))
-    assert eq(df.max(axis=1, skipna=False), ddf.max(axis=1, skipna=False))
-    assert eq(df.std(axis=1, skipna=False), ddf.std(axis=1, skipna=False))
-    assert eq(df.var(axis=1, skipna=False), ddf.var(axis=1, skipna=False))
-    assert eq(df.std(axis=1, skipna=False, ddof=0),
-              ddf.std(axis=1, skipna=False, ddof=0))
-    assert eq(df.var(axis=1, skipna=False, ddof=0),
-              ddf.var(axis=1, skipna=False, ddof=0))
-    assert eq(df.mean(axis=1, skipna=False), ddf.mean(axis=1, skipna=False))
-
-    assert eq(df.cumsum(), ddf.cumsum())
-    assert eq(df.cummin(), ddf.cummin())
-    assert eq(df.cummax(), ddf.cummax())
-    assert eq(df.cumprod(), ddf.cumprod())
-
-    assert eq(df.cumsum(skipna=False), ddf.cumsum(skipna=False))
-    assert eq(df.cummin(skipna=False), ddf.cummin(skipna=False))
-    assert eq(df.cummax(skipna=False), ddf.cummax(skipna=False))
-    assert eq(df.cumprod(skipna=False), ddf.cumprod(skipna=False))
-
-    assert eq(df.cumsum(axis=1), ddf.cumsum(axis=1))
-    assert eq(df.cummin(axis=1), ddf.cummin(axis=1))
-    assert eq(df.cummax(axis=1), ddf.cummax(axis=1))
-    assert eq(df.cumprod(axis=1), ddf.cumprod(axis=1))
-
-    assert eq(df.cumsum(axis=1, skipna=False), ddf.cumsum(axis=1, skipna=False))
-    assert eq(df.cummin(axis=1, skipna=False), ddf.cummin(axis=1, skipna=False))
-    assert eq(df.cummax(axis=1, skipna=False), ddf.cummax(axis=1, skipna=False))
-    assert eq(df.cumprod(axis=1, skipna=False), ddf.cumprod(axis=1, skipna=False))
 
 
 def test_describe():
@@ -1306,38 +368,6 @@ def test_drop_duplicates_subset():
                       ddf.drop_duplicates(subset=ss, **kwarg))
 
 
-def test_full_groupby():
-    assert raises(Exception, lambda: d.groupby('does_not_exist'))
-    assert raises(Exception, lambda: d.groupby('a').does_not_exist)
-    assert 'b' in dir(d.groupby('a'))
-    def func(df):
-        df['b'] = df.b - df.b.mean()
-        return df
-
-    assert eq(d.groupby('a').apply(func), full.groupby('a').apply(func))
-
-    assert sorted(d.groupby('a').apply(func).dask) == \
-           sorted(d.groupby('a').apply(func).dask)
-
-
-def test_groupby_on_index():
-    e = d.set_index('a')
-    efull = full.set_index('a')
-    assert eq(d.groupby('a').b.mean(), e.groupby(e.index).b.mean())
-
-    def func(df):
-        df.loc[:, 'b'] = df.b - df.b.mean()
-        return df
-
-    assert eq(d.groupby('a').apply(func).set_index('a'),
-              e.groupby(e.index).apply(func))
-    assert eq(d.groupby('a').apply(func), full.groupby('a').apply(func))
-    assert eq(d.groupby('a').apply(func).set_index('a'),
-              full.groupby('a').apply(func).set_index('a'))
-    assert eq(efull.groupby(efull.index).apply(func),
-              e.groupby(e.index).apply(func))
-
-
 def test_set_partition():
     d2 = d.set_partition('b', [0, 2, 9])
     assert d2.divisions == (0, 2, 9)
@@ -1391,27 +421,6 @@ def test_get_division():
 
     assert raises(ValueError, lambda: ddf.get_division(-1))
     assert raises(ValueError, lambda: ddf.get_division(3))
-
-def test_categorize():
-    dsk = {('x', 0): pd.DataFrame({'a': ['Alice', 'Bob', 'Alice'],
-                                   'b': ['C', 'D', 'E']},
-                                   index=[0, 1, 2]),
-           ('x', 1): pd.DataFrame({'a': ['Bob', 'Charlie', 'Charlie'],
-                                   'b': ['A', 'A', 'B']},
-                                   index=[3, 4, 5])}
-    d = dd.DataFrame(dsk, 'x', ['a', 'b'], [0, 3, 5])
-    full = d.compute()
-
-    c = d.categorize('a')
-    cfull = c.compute()
-    assert cfull.dtypes['a'] == 'category'
-    assert cfull.dtypes['b'] == 'O'
-
-    assert list(cfull.a.astype('O')) == list(full.a)
-
-    assert (d._get(c.dask, c._keys()[:1])[0].dtypes == cfull.dtypes).all()
-
-    assert (d.categorize().compute().dtypes == 'category').all()
 
 
 def test_ndim():
@@ -1942,59 +951,6 @@ def test_series_nunique():
     assert eq(s.nunique(), ps.nunique())
 
 
-def test_dataframe_groupby_nunique():
-    strings = list('aaabbccccdddeee')
-    data = np.random.randn(len(strings))
-    ps = pd.DataFrame(dict(strings=strings, data=data))
-    s = dd.from_pandas(ps, npartitions=3)
-    expected = ps.groupby('strings')['data'].nunique()
-    assert eq(s.groupby('strings')['data'].nunique(), expected)
-
-
-def test_dataframe_groupby_nunique_across_group_same_value():
-    strings = list('aaabbccccdddeee')
-    data = list(map(int, '123111223323412'))
-    ps = pd.DataFrame(dict(strings=strings, data=data))
-    s = dd.from_pandas(ps, npartitions=3)
-    expected = ps.groupby('strings')['data'].nunique()
-    assert eq(s.groupby('strings')['data'].nunique(), expected)
-
-
-@pytest.mark.parametrize(['npartitions', 'freq', 'closed', 'label'],
-                         list(product([2, 5], ['30T', 'h', 'd', 'w', 'M'],
-                                      ['right', 'left'], ['right', 'left'])))
-def test_series_resample(npartitions, freq, closed, label):
-    index = pd.date_range('1-1-2000', '2-15-2000', freq='h')
-    index = index.union(pd.date_range('4-15-2000', '5-15-2000', freq='h'))
-    df = pd.Series(range(len(index)), index=index)
-    ds = dd.from_pandas(df, npartitions=npartitions)
-    # Series output
-    result = ds.resample(freq, how='mean', closed=closed, label=label).compute()
-    expected = df.resample(freq, how='mean', closed=closed, label=label)
-    tm.assert_series_equal(result, expected, check_dtype=False)
-    # Frame output
-    resampled = ds.resample(freq, how='ohlc', closed=closed, label=label)
-    divisions = resampled.divisions
-    result = resampled.compute()
-    expected = df.resample(freq, how='ohlc', closed=closed, label=label)
-    tm.assert_frame_equal(result, expected, check_dtype=False)
-    assert expected.index[0] == divisions[0]
-    assert expected.index[-1] == divisions[-1]
-
-
-def test_series_resample_not_implemented():
-    index = pd.date_range(start='20120102', periods=100, freq='T')
-    s = pd.Series(range(len(index)), index=index)
-    ds = dd.from_pandas(s, npartitions=5)
-    # Frequency doesn't evenly divide day
-    assert raises(NotImplementedError, lambda: ds.resample('57T'))
-    # Kwargs not implemented
-    kwargs = {'fill_method': 'bfill', 'limit': 2, 'loffset': 2, 'base': 2,
-              'convention': 'end', 'kind': 'period'}
-    for k, v in kwargs.items():
-        assert raises(NotImplementedError, lambda: ds.resample('6h', **{k: v}))
-
-
 def test_set_partition_2():
     df = pd.DataFrame({'x': [1, 2, 3, 4, 5, 6], 'y': list('abdabd')})
     ddf = dd.from_pandas(df, 2)
@@ -2232,20 +1188,6 @@ def test_nlargest_series():
     assert eq(ss.nlargest(2), s.nlargest(2))
 
 
-def test_categorical_set_index():
-    df = pd.DataFrame({'x': [1, 2, 3, 4], 'y': ['a', 'b', 'b', 'c']})
-    df['y'] = df.y.astype('category')
-    a = dd.from_pandas(df, npartitions=2)
-
-    with dask.set_options(get=get_sync):
-        b = a.set_index('y')
-        df2 = df.set_index('y')
-        assert list(b.index.compute()), list(df2.index)
-
-        b = a.set_index(a.y)
-        df2 = df.set_index(df.y)
-        assert list(b.index.compute()), list(df2.index)
-
 
 def test_query():
     df = pd.DataFrame({'x': [1, 2, 3, 4], 'y': [5, 6, 7, 8]})
@@ -2348,41 +1290,6 @@ def test_to_frame():
     assert eq(s.to_frame(), a.to_frame())
     assert eq(s.to_frame('bar'), a.to_frame('bar'))
 
-
-def test_series_groupby_propagates_names():
-    df = pd.DataFrame({'x': [1, 2, 3], 'y': [4, 5, 6]})
-    ddf = dd.from_pandas(df, 2)
-    func = lambda df: df['y'].sum()
-
-    result = ddf.groupby('x').apply(func, columns='y')
-
-    expected = df.groupby('x').apply(func)
-    expected.name = 'y'
-
-    tm.assert_series_equal(result.compute(), expected)
-
-
-def test_series_groupby():
-    s = pd.Series([1, 2, 2, 1, 1])
-    pd_group = s.groupby(s)
-
-    ss = dd.from_pandas(s, npartitions=2)
-    dask_group = ss.groupby(ss)
-
-    pd_group2 = s.groupby(s + 1)
-    dask_group2 = ss.groupby(ss + 1)
-
-    for dg, pdg in [(dask_group, pd_group), (pd_group2, dask_group2)]:
-        assert eq(dg.count(), pdg.count())
-        assert eq(dg.sum(), pdg.sum())
-        assert eq(dg.min(), pdg.min())
-        assert eq(dg.max(), pdg.max())
-
-    assert raises(TypeError, lambda: ss.groupby([1, 2]))
-    sss = dd.from_pandas(s, npartitions=3)
-    assert raises(NotImplementedError, lambda: ss.groupby(sss))
-
-
 def test_apply():
     df = pd.DataFrame({'x': [1, 2, 3, 4], 'y': [10, 20, 30, 40]})
     a = dd.from_pandas(df, npartitions=2)
@@ -2433,21 +1340,6 @@ def test_nlargest_multiple_columns():
     result = ddf.nlargest(5, ['a', 'b'])
     expected = df.nlargest(5, ['a', 'b'])
     eq(result, expected)
-
-
-def test_groupby_index_array():
-    df = tm.makeTimeDataFrame()
-    ddf = dd.from_pandas(df, npartitions=2)
-
-    eq(df.A.groupby(df.index.month).nunique(),
-       ddf.A.groupby(ddf.index.month).nunique(), check_names=False)
-
-
-def test_groupby_set_index():
-    df = tm.makeTimeDataFrame()
-    ddf = dd.from_pandas(df, npartitions=2)
-    assert raises(NotImplementedError,
-                  lambda: ddf.groupby(df.index.month, as_index=False))
 
 
 def test_reset_index():
@@ -2511,12 +1403,3 @@ def test_to_imperative():
     assert isinstance(b, Value)
 
     assert eq(a.compute(), df.iloc[:2])
-
-
-def test_dataframe_categoricals():
-    df = pd.DataFrame({'x': list('a'*5 + 'b'*5 + 'c'*5),
-                       'y': range(15)})
-    df.x = df.x.astype('category')
-    ddf = dd.from_pandas(df, npartitions=2)
-    assert (df.x.cat.categories == pd.Index(['a', 'b', 'c'])).all()
-    assert not hasattr(df.y, 'cat')
