+++ dask/dataframe/utils.py
@@ -47,6 +47,8 @@ def shard_df_on_index(df, divisions):
     3  30  2
     4  40  1
     """
+    from dask.dataframe.categorical import iscategorical
+
     if isinstance(divisions, Iterator):
         divisions = list(divisions)
     if not len(divisions):
@@ -85,86 +87,6 @@ def unique(divisions):
     if isinstance(divisions, (tuple, list, Iterator)):
         return tuple(toolz.unique(divisions))
     raise NotImplementedError()
-
-
-def _categorize(categories, df):
-    """ Categorize columns in dataframe
-
-    >>> df = pd.DataFrame({'x': [1, 2, 3], 'y': [0, 2, 0]})
-    >>> categories = {'y': ['A', 'B', 'c']}
-    >>> _categorize(categories, df)
-       x  y
-    0  1  A
-    1  2  c
-    2  3  A
-
-    >>> _categorize(categories, df.y)
-    0    A
-    1    c
-    2    A
-    dtype: category
-    Categories (3, object): [A, B, c]
-    """
-    if '.index' in categories:
-        index = pd.CategoricalIndex(
-                  pd.Categorical.from_codes(df.index.values, categories['.index']))
-    else:
-        index = df.index
-    if isinstance(df, pd.Series):
-        if df.name in categories:
-            cat = pd.Categorical.from_codes(df.values, categories[df.name])
-            return pd.Series(cat, index=index)
-        else:
-            return df
-
-    else:
-        return pd.DataFrame(
-                dict((col, pd.Categorical.from_codes(df[col].values, categories[col])
-                           if col in categories
-                           else df[col].values)
-                    for col in df.columns),
-                columns=df.columns,
-                index=index)
-
-
-def strip_categories(df):
-    """ Strip categories from dataframe
-
-    >>> df = pd.DataFrame({'x': [1, 2, 3], 'y': ['A', 'B', 'A']})
-    >>> df['y'] = df.y.astype('category')
-    >>> strip_categories(df)
-       x  y
-    0  1  0
-    1  2  1
-    2  3  0
-    """
-    return pd.DataFrame(dict((col, df[col].cat.codes.values
-                                   if iscategorical(df.dtypes[col])
-                                   else df[col].values)
-                              for col in df.columns),
-                        columns=df.columns,
-                        index=df.index.codes
-                              if iscategorical(df.index.dtype)
-                              else df.index)
-
-def iscategorical(dt):
-    return isinstance(dt, pd.core.common.CategoricalDtype)
-
-
-def get_categories(df):
-    """
-    Get Categories of dataframe
-
-    >>> df = pd.DataFrame({'x': [1, 2, 3], 'y': ['A', 'B', 'A']})
-    >>> df['y'] = df.y.astype('category')
-    >>> get_categories(df)
-    {'y': Index([u'A', u'B'], dtype='object')}
-    """
-    result = dict((col, df[col].cat.categories) for col in df.columns
-                  if iscategorical(df.dtypes[col]))
-    if iscategorical(df.index.dtype):
-        result['.index'] = df.index.categories
-    return result
 
 
 ###############################################################
